+++ templates/workspaces/base/terraform/airlock/outputs.tf
@@ -1,0 +1,19 @@
+output "import_approved_storage_name" {
+  value = local.import_approved_storage_name
+}
+
+output "export_internal_storage_name" {
+  value = local.export_internal_storage_name
+}
+
+output "export_inprogress_storage_name" {
+  value = local.export_inprogress_storage_name
+}
+
+output "export_rejected_storage_name" {
+  value = local.export_rejected_storage_name
+}
+
+output "export_blocked_storage_name" {
+  value = local.export_blocked_storage_name
+}